Add debug output to search function, add helper script to set up DB after clearing it

+++ crates/libs/app_state/src/raw_settings.rs
@@ -35,7 +35,7 @@ pub struct AnalyzerSettings {
 pub struct SearchSettings {
     pub embedder_model_id: String,
     pub semantic_weight: f64,
-    pub text_weight: f64,
+    pub text_weight: f32,
 }
 
 #[derive(Debug, Deserialize, Clone)]
